rename: UserSearchResultBox -> UserSearchResultBox
